change all file handling classes and add logging to application
