+++ src/main/java/accounts/CurrentAccount.java
@@ -2,7 +2,7 @@ package accounts;
 
 import bank.Bank;
 import currencies.CurrencyCodes;
-import file_manipulation.TransactionHistoryToCSV;
+import file_manipulation.TransactionHistoryCSVHandler;
 import transaction.Transaction;
 import transaction.TransactionTypes;
 import users.User;
@@ -63,7 +63,7 @@ public class CurrentAccount extends Account {
             super.getTransactionHistory().add(new Transaction(getAccountNumber(), TransactionTypes.WITHDRAW, LocalDateTime.now(), amount, getCurrencyCode()));
             // Checks if the account exist in bank. Accounts created by unit testing are not included, so we don't need to save their transaction history.
             if (Bank.getInstance().contains(getAccountNumber())) {
-                TransactionHistoryToCSV.write(getTransactionHistory(), new File("transactions/transaction_history_" + getAccountNumber() + ".csv").getAbsolutePath());
+                TransactionHistoryCSVHandler.write(getTransactionHistory(), new File("transactions/transaction_history_" + getAccountNumber() + ".csv").getAbsolutePath());
             }
             return true;
         }
