+++ src/main/java/file_manipulation/FileManipulator.java
@@ -6,14 +6,14 @@ import java.io.File;
 
 public class FileManipulator {
     public static void loadDataFromFile() {
-        CSVToUsers.read(Bank.getInstance(), new File("users.csv").getAbsolutePath());
-        CSVToAccounts.read(Bank.getInstance(), new File("accounts.csv").getAbsolutePath());
-        CSVToAccountNumber.read(Bank.getInstance().getAccountNumbers(), new File("account_numbers.csv").getAbsolutePath());
+        UsersCSVHandler.read(Bank.getInstance(), new File("users.csv").getAbsolutePath());
+        AccountsCSVHandler.read(Bank.getInstance(), new File("accounts.csv").getAbsolutePath());
+        AccountNumberCSVHandler.read(Bank.getInstance().getAccountNumbers(), new File("account_numbers.csv").getAbsolutePath());
     }
 
     public static void saveDataToFile() {
-        UsersToCSV.write(Bank.getInstance().getAllUsers(), new File("users.csv").getAbsolutePath());
-        AccountsToCSV.write(Bank.getInstance().getAllAccounts(), new File("accounts.csv").getAbsolutePath());
-        AccountNumberToCSV.write(Bank.getInstance().getAccountNumbers(), new File("account_numbers.csv").getAbsolutePath());
+        UsersCSVHandler.write(Bank.getInstance().getAllUsers(), new File("users.csv").getAbsolutePath());
+        AccountsCSVHandler.write(Bank.getInstance().getAllAccounts(), new File("accounts.csv").getAbsolutePath());
+        AccountNumberCSVHandler.write(Bank.getInstance().getAccountNumbers(), new File("account_numbers.csv").getAbsolutePath());
     }
 }
