+++ src/main/java/file_manipulation/LogoLoader.java
@@ -2,8 +2,11 @@ package file_manipulation;
 
 import java.io.*;
 import java.util.Objects;
+import java.util.logging.Logger;
 
 public class LogoLoader {
+    private static final Logger LOGGER = Logger.getLogger(AccountNumberCSVHandler.class.getName());
+
     public static String read(String fileName) {
         StringBuilder resultStringBuilder = new StringBuilder();
 
@@ -14,8 +17,10 @@ public class LogoLoader {
                 resultStringBuilder.append(line).append("\n");
             }
 
+            LOGGER.info("Logo successfully loaded from " + fileName);
             return resultStringBuilder.toString();
         } catch (IOException e) {
+            LOGGER.severe("Failed to load logo from " + fileName + ": " + e.getMessage());
             throw new RuntimeException(e.getMessage());
         }
     }
