+++ src/main/resources/logging.properties
@@ -1,0 +1,12 @@
+handlers = java.util.logging.ConsoleHandler, java.util.logging.FileHandler
+.level = INFO
+
+java.util.logging.ConsoleHandler.level = INFO
+java.util.logging.ConsoleHandler.formatter = java.util.logging.SimpleFormatter
+
+java.util.logging.FileHandler.level = INFO
+java.util.logging.FileHandler.pattern = logs/app.log
+java.util.logging.FileHandler.limit = 50000
+java.util.logging.FileHandler.count = 1
+java.util.logging.FileHandler.append = true
+java.util.logging.FileHandler.formatter = java.util.logging.SimpleFormatter
